refactor:views폴더 이름 view로 변경
