+++ src/controller/RacingGame.js
@@ -1,6 +1,6 @@
 const Car = require("../domain/Car");
-const InputView = require("../views/InputView");
-const OutputView = require("../views/OutputView");
+const InputView = require("../view/InputView");
+const OutputView = require("../view/OutputView");
 const RandomNumberGenerator = require("../utils/RandomNumberGenerator")
 class RacingGame {
   #cars;
